requirements.txt
appdirs==1.4.4
black==20.8b1
click==7.1.2
mypy-extensions==0.4.3
mysql-connector-python==8.0.23
numpy==1.20.1
pandas==1.2.3
pathspec==0.8.1
protobuf==3.15.6
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
regex==2021.3.17
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
XlsxWriter==1.3.7
